pyproject.toml
[project]
name = "cdpmd"
version = "0.1.0"
description = "Chronic Disease Progressive Model for Diabetes (CDPMD)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "authlib>=1.5.0",
    "fhir-resources>=8.0.0",
    "httpx[http2]>=0.28.1",
    "logfire[httpx]>=3.6.4",
    "pydantic-ai>=0.0.26",
    "python-dotenv>=1.0.1",
    "python-fasthtml>=0.12.1",
]
